Organizando HomePage creado el directorio para sus componentes, añadidos los componentes header, navBar y Main, larga vida y prosperidad
