+++ RetroFrog V3/src/pages/HomeComponents/MainPage.tsx
@@ -1,6 +1,6 @@
 
 
-export const Inicio=()=>{
+export const HomeInicio=()=>{
     
     return <main className="">Hola mundo</main>
  }
